+++ src/app/modules/finance/finance.component.ts
@@ -1,5 +1,5 @@
 import { COMMA, ENTER } from '@angular/cdk/keycodes';
-import { registerLocaleData } from '@angular/common';
+import { DecimalPipe, registerLocaleData } from '@angular/common';
 import es from '@angular/common/locales/es';
 import { Component, ElementRef, OnInit, ViewChild } from '@angular/core';
 import { FormControl } from '@angular/forms';
@@ -23,23 +23,34 @@ registerLocaleData(es, 'es');
   ]
 })
 export class FinanceComponent implements OnInit {
-
-  visible = true;
+  /** True if a tag can be selected */
   selectable = true;
+  /** True if can remove a tag */
   removable = true;
+  /** separators */
   separatorKeysCodes: number[] = [ENTER, COMMA];
+  /** Tag form control */
   tagCtrl = new FormControl();
+  /** Filtered tags array */
   filteredTags: Observable<string[]>;
+  /** Tags */
   tags: string[] = [];
-  allTags: string[] = ['Apple', 'Lemon', 'Lime', 'Orange', 'Strawberry'];
+  /** Al tags */
+  allTags: string[] = [];
+  /** Column names array  */
   displayedColumns: string[] = ['date', 'name', 'total'];
 
+  /** Tags input control  */
   @ViewChild('tagInput') tagInput: ElementRef<HTMLInputElement>;
+  /** Autocomplete component */
   @ViewChild('auto') matAutocomplete: MatAutocomplete;
 
+  /** Material accordion component */
   @ViewChild(MatAccordion) accordion: MatAccordion;
   
+  /** Table data Source */
   tableDataSource: MatTableDataSource<any>;
+  /** Material paginator */
   @ViewChild(MatPaginator) paginator: MatPaginator;
 
   selectedGroupDate: string = 'day';
@@ -80,19 +91,31 @@ export class FinanceComponent implements OnInit {
   public lineChartType = 'line';
   public lineChartPlugins = [];
   
+  /**
+   * Constructor
+   * @param financeService 
+   * @param _decimalPipe 
+   */
   constructor(
-    private financeService: FinanceService
+    private financeService: FinanceService,
+    private _decimalPipe: DecimalPipe
   ) {
     this.financeService.getStats();
     this.filterTags();
    }
 
+   /**
+    * Filter list tags
+    */
    private filterTags() {
     this.filteredTags = this.tagCtrl.valueChanges.pipe(
       startWith(''),
       map((tag: string | null) => tag ? this._filter(tag) : this.allTags.filter(tag => !this.tags.includes(tag))));
    }
 
+  /**
+   * On init
+   */
   ngOnInit(): void {
     this.groupData();
   }
@@ -130,7 +153,9 @@ export class FinanceComponent implements OnInit {
         this.globalTotal = 0;
         for( const key in totalBy[this.selectedGroupDate]) {
           this.globalTotal += totalBy[this.selectedGroupDate][key];
-          data.push(totalBy[this.selectedGroupDate][key]);
+          let total: number = totalBy[this.selectedGroupDate][key];
+          total = Math.round(total * 100) / 100;
+          data.push(total);
           this.lineChartLabels.push(key);
         }
         if (data.length > 0 ) this.lineChartData.push({ data })
@@ -139,7 +164,10 @@ export class FinanceComponent implements OnInit {
       }
     );    
   }
-
+  /**
+   * Set table data source
+   * @param data 
+   */
   public setTableDataSource(data) {
     const datasource = data.map( item => {
       return {